Fix buffer overflow in File::path()
In netcdf-c, nc_inq_path() uses strlen() to calculate the length
of the string [1]. strlen() does not include the trailing NUL in
its calculations… but the strcpy() does, and it includes one.

This makes the buffer allocated in File::path() always one byte
too small to receive the path string. The result is a buffer
overflow and undefined behavior. The presence of memory safety
issues can be confirmed with valgrind.

Fix by adding a byte to the size returned by C. The NUL byte must
be truncated off before the Vec is converted to String.

References
1. https://github.com/Unidata/netcdf-c/blob/v4.7.4/libdispatch/dfile.c#L889